104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,16 +5,16 @@
  */
 int main(void)
 {
-	unsigned int n1 = 1, n2 = 2, n3;
+	long unsigned int n1 = 1, n2 = 2, n3;
 	int i;
 
-	printf("%u, %u, ", n1, n2);
+	printf("%lu, %lu, ", n1, n2);
 		for (i = 1; i <= 96; i++)
 		{
 			n3 = n1 + n2;
 			n1 = n2;
 			n2 = n3;
-			printf("%u", n3);
+			printf("%lu", n3);
 			if (i != 48)
 			{
 				printf(", ");
